refactor: Domain과 연관된 rules, errorMessage 상수들 Domain 폴더로 이동
