+++ src/Domain/Constants/errorMessage.js
@@ -3,7 +3,7 @@ import { MAX, MIN } from './rules.js';
 const ERROR_PREFIX = '[Error]';
 
 export const CAR_NAMES = {
-  LENGTH: `${ERROR_PREFIX} 자동차 이름은 ${MAX.CAR_NAME_LENGTH}자 초과 안됩니다.`,
+  LENGTH: `${ERROR_PREFIX} 자동차 이름은 ${MIN.CAR_NAME_LENGTH}자 미만, ${MAX.CAR_NAME_LENGTH}자 초과 안됩니다.`,
   FORM: `${ERROR_PREFIX} 자동차 이름이 올바르지 않습니다.`,
   DUPLICATED: `${ERROR_PREFIX} 중복된 자동차 이름은 안됩니다.`,
 };
